// src/app/models/AddOnsContent.ts
export interface AddOnsContent {
    subscriptionSection: {
        title: string;
        highlight: string;
        description: string;
    };
    addonsSection: {
        title: string;
        description: string;
    };
    buttonText: {
        showMore: string;
        showLess: string;
    };
    styling: {
        hrWidth: string;
    };
}